FastApi/sima/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
attrs==23.1.0
bcrypt==4.1.1
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
ecdsa==0.18.0
elastic-transport==8.10.0
elasticsearch==8.11.1
et-xmlfile==1.1.0
fastapi==0.105.0
fastapi-cache2==0.2.1
fastapi-pagination==0.12.13
frozenlist==1.4.0
h11==0.14.0
idna==3.6
Jinja2==3.1.2
loguru==0.7.2
MarkupSafe==2.1.3
multidict==6.0.4
openpyxl==3.1.2
orjson==3.9.10
passlib==1.7.4
pendulum==2.1.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyMySQL==1.1.0
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.24.0.post1
XlsxWriter==3.1.9
yarl==1.9.4
